Add quiz questions to pulled quizzes
It's kind of messy, but it works. The worst part is the QuizAnswers and
QuizAnswerMatches nested structs. I gave up and changed them to maps for
now. Implementing the unmarshaler interface is basically done though, I
think I just need to check for nil values before performing the type
conversions.